+++ lib/warc.dart
@@ -1,4 +1,5 @@
 export 'src/cdxj_record.dart';
+export 'src/cdxj_writer.dart';
 export 'src/warc_record.dart';
 export 'src/warc_reader.dart';
 export 'src/warc_writer.dart';